// src/types/error-codes.ts
export enum CODE {
  UNAUTHORIZED = "UNAUTHORIZED",
  BAD_REQUEST = "BAD_REQUEST",
  UNEXPECTED_ERROR = "UNEXPECTED_ERROR",
  NOT_FOUND = "NOT_FOUND",
  TOKEN_EXPIRED = "TOKEN_EXPIRED",
  VALIDATION_REQUEST_ERROR = "VALIDATION_REQUEST_ERROR",
  EMAIL_IN_USE = "EMAIL_IN_USE",
  ACCESS_TOKEN_EXPIRED = "ACCESS_TOKEN_EXPIRED",
  REFRESH_TOKEN_EXPIRED = "REFRESH_TOKEN_EXPIRED",
  REUSE_DETECTION = "REUSE_DETECTION",
  NO_ACCESS_TOKEN = "NO_ACCESS_TOKEN",
  NO_REFRESH_TOKEN = "NO_REFRESH_TOKEN",
  UNABLE_TO_LOGIN = "UNABLE_TO_LOGIN",
  MULTER_VIDEO_TYPE_ERROR = "MULTER_VIDEO_TYPE_ERROR",
  MULTER_SIZE_ERROR = "MULTER_SIZE_ERROR",
  MULTER_FILE_DOES_NOT_EXIST = "MULTER_FILE_DOES_NOT_EXIST",
  VIDEO_CONVERTED_ALREADY = "VIDEO_CONVERTED_ALREADY",
  VIDEO_IS_STILL_BEING_CONVERTED = "VIDEO_IS_STILL_BEING_CONVERTED",
  PASSWORD_DOES_NOT_MATCH = "PASSWORD_DOES_NOT_MATCH",
  PASSWORDS_MUST_BE_THE_SAME = "PASSWORDS_MUST_BE_THE_SAME",
}

export enum SUCCESS_CODE {
  SUCCESS = "SUCCESS",
}